NFDIV-770: Clean up cucumber fixtures
