+++ src/test/functional/fixtures/jointApplicant1CompleteCase.ts
@@ -51,8 +51,3 @@ export const jointApplicant1CompleteCase: Partial<BrowserCase> = {
   applicant1UploadedFiles: [],
   whoIsFinancialOrderFor: [],
 };
-
-export const jointApplicant1CompleteCaseHwf: Partial<BrowserCase> = {
-  ...jointApplicant1CompleteCase,
-  applicant1AlreadyAppliedForHelpPaying: YesOrNo.YES,
-};
